+++ src/components/shared/Graph.scss
